backend/requirements.txt
# requirements.txt — зависимости бэкенда
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.0.0
